Submission. I had meant to make a larger number of commits instead of a big one, but a lot of the time was spent trying to reorient myself with the tool set and ide, which is why it probably took longer than it should have. .NET Framework Core has a lot less "magic" than Spring Boot.
I decided to go with a Controller -> Repository -> DbContext pattern, as that would enable dependency injection and in a real app, you would not have the controller directly access a DbContext. In a real-world app, you'd probably make the IRepository even more generic than I have.

I'm not all that familiar with the C# testing utilities so I decided to skip wrestling with them for now and used Postman to test it out locally. The responses are as expected.

The main thing I don't like is the implementation of Update function. It's clunky in the way it tries to not accidentally blank out data.